% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranI.R
\name{calcMoran}
\alias{calcMoran}
\title{Moran I's test on one plate}
\usage{
calcMoran(
  onePlateData,
  targetColumn = "CalcResultI.RAW_VALUE",
  distMethod = "euclidean"
)
}
\arguments{
\item{onePlateData}{data for one plate containing rows and columns for a specific well type.
Note that subsetting plates into negative/positive controls should be done outside this function.}

\item{targetColumn}{character, the column with the raw data}

\item{distMethod}{string indicating which distance metric to be used.
Can be "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski" with euclidean as default.}
}
\value{
returns the p-value of the null hypothesis 
that there is zero spatial autocorrelation present in targetColumn
}
\description{
Moran’s I is a measure of spatial autocorrelation.
}
\author{
hwouters (adapted, hbossier)
}
