% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMoranforallWELLTYPES.R
\name{calcMoranforallWELLTYPES}
\alias{calcMoranforallWELLTYPES}
\title{Calculate Moran's I for all Well Types}
\usage{
calcMoranforallWELLTYPES(plateData, featureName, ...)
}
\arguments{
\item{plateData}{data for one plate containing rows and columns with negative controls}

\item{featureName}{character, the column with the data to test}

\item{...}{additional arguments}
}
\value{
returns vector of p-value for Moran's I autocorrelation test 
for the different welltypes in WELLTYPE_CODE and the combination of NC and SAMPLE
}
\description{
Calculate Moran's I for all Well Types
}
\author{
hwouters
}
