% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIntraPlateEffect.R
\name{checkIntraPlateEffect}
\alias{checkIntraPlateEffect}
\title{Check Intra Plate effects}
\usage{
checkIntraPlateEffect(
  plateData,
  featureName = "RAW_DATA",
  adjustmentMethod = "fdr"
)
}
\arguments{
\item{plateData}{plate data with columns \code{WELLTYPE_CODE}, \code{ROW_NR}, 
\code{COL_NR} and the feature specified in the \code{featureName} argument.}

\item{featureName}{variable to be tested for intra-plate effect}

\item{adjustmentMethod}{adjustment for multiplicity; default value \code{"fdr"}}
}
\value{
list of results: significant rows/columns and number of significant 
rows and colums for each WELLTYPE_CODE S,NC,P andSNC
}
\description{
Check Intra Plate effect by test that compares each of the rows with all other 
rows and each of the columns with all other columns.
}
