% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcolMWtest.R
\name{colMWtest}
\alias{colMWtest}
\title{For a column calculate the Man Withney U test statistic comparing the row with all other rows}
\usage{
colMWtest(columnNumber, plateData, featureName = "RAW_DATA")
}
\arguments{
\item{columnNumber}{column number}

\item{plateData}{plate data}

\item{featureName}{variable to be tested for intra-plate effect; default value \code{"RAW_DATA"}}
}
\value{
Man Withney U test statistic p-values
}
\description{
For a column calculate the Man Withney U test statistic comparing the row with all other rows
}
