% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBScore.R
\name{computeBScore}
\alias{computeBScore}
\title{Compute a B Score}
\usage{
computeBScore(input, correctionType)
}
\arguments{
\item{input}{input data}

\item{correctionType}{correction type; should be one of the values of the \code{WELLTYPE_CODE} column;
default value \code{"NC"}}
}
\description{
Compute a B Score
}
