% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePes.R
\name{computePes}
\alias{computePes}
\title{Compute the PES}
\usage{
computePes(input)
}
\arguments{
\item{input}{input data frame; it is assumed the columns 
\code{WELLTYPE_CODE} and \code{correction} are present}
}
\value{
data frame with input data and additional columns \code{zFactor} and \code{pes}
}
\description{
Compute the PES
}
