% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRobustZ.R
\name{computeRobustZ}
\alias{computeRobustZ}
\title{Compute the Robust Z Score}
\usage{
computeRobustZ(input, correctionType)
}
\arguments{
\item{input}{input data}

\item{correctionType}{correction type; should be one of the values of the \code{WELLTYPE_CODE} column;
default value \code{"NC"}}
}
\description{
Compute the Robust Z Score
}
