% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlate.R
\name{createPlate}
\alias{createPlate}
\title{Create matrix with data from specific plate - long format to plate format}
\usage{
createPlate(plateData, rowName, columnName, plateName = NULL, valueName)
}
\arguments{
\item{plateData}{data.frame, data from specific plate; should at least include variable names specified in \code{c(rowName, columnName, valueName)}}

\item{rowName}{character, variable name in \code{plateData} that specifies the plate row of each data point}

\item{columnName}{character, variable name in \code{plateData} that specifies the plate row of each data point}

\item{plateName}{character, variable name in \code{plateData} that specifies the plate name}

\item{valueName}{character, variable name in \code{plateData} that specifies the value of each data point}
}
\value{
matrix, with data from \code{plateData} in long format to plate format
}
\description{
Create matrix with data from specific plate - long format to plate format
}
\author{
hwouters reused code from mvarewyck
}
