% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessProcedure.R
\name{loessProcedure}
\alias{loessProcedure}
\title{Runs Loess correction}
\usage{
loessProcedure(
  plateData,
  moranIPthresh = 0.05,
  featureName,
  correction_type = "NC",
  span = 0.9,
  degree = 1,
  ...
)
}
\arguments{
\item{plateData}{plate data containing rows and columns with negative controls}

\item{moranIPthresh}{threshold to determine statistical significance of the the Moran I test.
This is only used to flag plates as a warning message (default = 0.05).}

\item{featureName}{Name of the measurement column (single feature)}

\item{correction_type}{default \code{"NC"} based on entries from the 
WELLTYPE_CODE column. Character vector with 1 or several of 
NC, PC or SAMPLE entries.
The specified WELLTYPE_CODE type should be distributed all over the plate.
This is needed for e.g. rescreens with biased samples.}

\item{span}{parameter for loess (default = 0.9), see \code{help(loess)}}

\item{degree}{parameter for loess (default = 1)}

\item{...}{parameters passed to \code{runMoranPlates()}.}
}
\value{
a data frame containing input data with addition of corrected columns
}
\description{
This function will fit a local polynomial regression model. Predicted values are 
calculated and subtracted from RAW_VALUES. Then the median value of the RAW_VALUES
is added to obtain loess corrected values for the entire plate.
}
\details{
The following model is fit:
loess(RAW_DATA ~ Row + Column + Row * Column,
neg_control_data, family="symmetric",
control=loess.control(surface="direct"), ...)
}
\author{
SWink, HBossier
}
