% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayCorrector.R
\name{print.assay}
\alias{print.assay}
\title{Print assay summary}
\usage{
\method{print}{assay}(x, ..., plate = 1)
}
\arguments{
\item{x}{The assay you want to print}

\item{...}{Ellipsis to be passed to the default \code{print()} function}

\item{plate}{The plate number (Default:1)}
}
\value{
None
}
\description{
\code{print.assay} simply prints a summary of the HTS assay
}
