% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcolMWtest.R
\name{rowMWtest}
\alias{rowMWtest}
\title{For a row calculate the Man Withney U test statistic comparing the row with all other rows}
\usage{
rowMWtest(rowNumber, plateData, featureName = "RAW_DATA")
}
\arguments{
\item{rowNumber}{row number}

\item{plateData}{plate data}

\item{featureName}{variable to be tested for intra-plate effect}
}
\value{
Man Withney U test statistic p-values
}
\description{
For a row calculate the Man Withney U test statistic comparing the row with all other rows
}
