% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessProcedure.R
\name{run_loess}
\alias{run_loess}
\title{Calculate loess plate correction}
\usage{
run_loess(input, featureName, correction_type, span = 0.9, degree = 1)
}
\arguments{
\item{input}{one plate dataframe}

\item{featureName}{Name of the measurement column (single feature)}

\item{correction_type}{default \code{"NC"} based on entries from the 
WELLTYPE_CODE column. Character vector with 1 or several of 
NC, PC or SAMPLE entries.
The specified WELLTYPE_CODE type should be distributed all over the plate.
This is needed for e.g. rescreens with biased samples.}

\item{span}{parameter for loess (default = 0.9), see \code{help(loess)}}

\item{degree}{parameter for loess (default = 1)}
}
\value{
the input data frame with additional columns: the corrected values and 
an identifier to denote the type of correction ('loess')
}
\description{
Calculates residuals of a loess plate correction


It requires the presence of the columns 'COL_NR' and 'ROW_NR', and the
column 'WELLTYPE_CODE', as well as the column with name RAW_VALUE 
which contains the measured data.
}
\author{
Steven Wink, adapted by Han Bossier
}
