% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PMP.R
\name{run_PMP}
\alias{run_PMP}
\title{Calculate new variant of PMP plate correction}
\usage{
run_PMP(plateData, featureName, correctionType)
}
\arguments{
\item{plateData}{data frame with columns WELLTYPE_CODE, ROW_NR, COL_NR and the feature specified in the featureName argument}

\item{featureName}{Name of the measurement column (single feature)}

\item{correctionType}{correction type; should be one of the values of the WELLTYPE_CODE column or, alternatively, "fullPlateRestoreMedians"; default value "NC"
The specified WELLTYPE_CODE type should be distributed all over the plate.
This is needed for e.g. rescreens with biased samples.}
}
\value{
the input data frame with additional columns: the corrected values and 
an identifier to denote the type of correction ('PMP')
}
\description{
Calculate new variant of PMP plate correction
It requires the presence of the columns 'COL_NR' and 'ROW_NR', and the
column 'WELLTYPE_CODE', as well as the column with name RAW_VALUE 
which contains the measured data.
}
\author{
Heidi Wouters
}
