% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranI.R
\name{runMoranPlates}
\alias{runMoranPlates}
\title{Moran I's test on multiple plates}
\usage{
runMoranPlates(
  plateData,
  targetColumn = "CalcResultI.RAW_VALUE",
  distMethod = "euclidean",
  splitColumn
)
}
\arguments{
\item{plateData}{plate data containing rows and columns with negative controls}

\item{targetColumn}{character, the column with the raw data}

\item{distMethod}{string indicating which distance metric to be used.
Can be "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski" with euclidean as default.}

\item{splitColumn}{character, the column with the plate IDs (where to split)}
}
\value{
returns the p-values of the null hypotheses
that there is zero spatial autocorrelation present in targetColumn for each plate
}
\description{
Moran’s I is a measure of spatial autocorrelation.
}
\author{
hbossier
}
